refactor(web): extract components from agents-showcase, onboarding, landing pages
