+++ apps/web/src/components/agents-showcase/index.ts
@@ -1,0 +1,7 @@
+export {
+  useAgentData,
+  OrchestratorHero,
+  PipelineSection,
+  AgentCard,
+} from './components';
+export type { AgentShowcase } from './components';
