+++ apps/web/src/components/landing/index.ts
@@ -1,0 +1,11 @@
+export {
+  SparklineChart,
+  TradeRow,
+  AgentDashboardCard,
+  FeatureCard,
+  ProviderCard,
+  AgentProfileCard,
+  ModeCard,
+  ProcessStep,
+} from './components';
+export type { FeatureCardProps } from './components';
