+++ apps/web/src/components/onboarding/index.ts
@@ -1,0 +1,11 @@
+export {
+  StepBinance,
+  StepLLM,
+  StepMode,
+  LLM_PROVIDERS,
+} from './components';
+export type {
+  TradingMode,
+  LLMProvider,
+  OnboardingState,
+} from './components';
